pyproject.toml
[tool.poetry]
name = "geotrouvetout"
version = "1.0.0"
description = "geotrouvetout gives geolocation information from an image"
authors = [
  "David Bret",
  "Paul Chambaz <paul.chambaz@tutanota.com>",
  "Feriel Cheggour",
  "Marion Mazaud"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pathlib = "^1.0.1"
fastapi = "^0.92.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
numpy = "^1.24.2"
python-multipart = "^0.0.5"
coloredlogs = "^15.0.1"
pillow = "^9.4.0"
requests = "^2.28.2"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl"}
torchvision = {url ="https://download.pytorch.org/whl/cpu/torchvision-0.14.1%2Bcpu-cp310-cp310-linux_x86_64.whl"}
pytesseract = "^0.3.10"
matplotlib = "^3.7.0"
ultralytics = "^8.0.40"
tesseract = "^0.1.3"
opencv-python = "^4.7.0.72"
langdetect = "^1.0.9"
collection = "^0.1.6"
pycountry = "^22.3.5"
nominatim = "^0.1"
countryinfo = "^0.1.2"
geopy = "^2.3.0"
csvfile = "^3.4.0"
googlemaps = "^4.10.0"
osmnx = "^1.3.0"
shapely = "^2.0.1"
scikit-image = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-clarity = "^1.0.1"
pytest-xdist = "^3.2.0"
mypy = "^1.0.0"
pylint = "^2.16.2"
pydocstyle = "^6.3.0"
pycodestyle = "^2.10.0"
black = "^23.1.0"
tox = "^4.4.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
geotrouvetout = "geotrouvetout:main"
